requirements.txt
# === Core Framework ===
fastapi==0.115.0
uvicorn==0.38.0
gradio==4.44.0

# === Web + Crawling + Parsing ===
feedparser==6.0.11
readability-lxml==0.8.1
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.1

# === Media Download / Processing ===
pytube==15.0.0
yt-dlp==2024.11.4
youtube-transcript-api==0.6.2
opencv-python-headless==4.10.0.84
moviepy==2.2.1
imageio==2.34.0
imageio-ffmpeg==0.5.1
pdfkit==1.0.0

# === AI / NLP / Embeddings ===
torch==2.5.1
transformers==4.45.2
sentence-transformers==2.7.0
underthesea==6.8.0
tqdm==4.66.5
numpy==2.1.2
scikit-learn==1.5.2

# === Data / Utilities ===
python-docx==1.1.2
pandas==2.2.2

# === Optional Database ===
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
